+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 * Problems when we stack multiple evaluations (see FIXME on repl-tooling)
 * Fix problems with unreadable forms
 
+## 0.1.2
+- Auto-adding port number if the project file is a shadow-cljs build project
+- Config to open console on bottom
+- Fixed console errors (https://github.com/mauricioszabo/atom-chlorine/issues/33) and (https://github.com/mauricioszabo/atom-chlorine/issues/32)
+
 ## 0.1.1
 - Fix `#js` tagged literal
 - Shadow-CLJS can now evaluate multiple forms